Move ResponseRenderer to pkg/lib/web
