+++ pkg/auth/deps.go
@@ -84,6 +84,7 @@ var DependencySet = wire.NewSet(
 	wire.Bind(new(oauthhandler.CookieManager), new(*httputil.CookieManager)),
 	wire.Bind(new(oauth.AppSessionTokenServiceCookieManager), new(*httputil.CookieManager)),
 	wire.Bind(new(handlerwebapp.TutorialCookie), new(*httputil.TutorialCookie)),
+	wire.Bind(new(handlerwebapp.Renderer), new(*web.ResponseRenderer)),
 	wire.Bind(new(handlerapi.WorkflowNewCookieManager), new(*httputil.CookieManager)),
 	wire.Bind(new(handlerapi.WorkflowInputCookieManager), new(*httputil.CookieManager)),
 	wire.Bind(new(handlerapi.WorkflowGetCookieManager), new(*httputil.CookieManager)),
@@ -136,7 +137,7 @@ var DependencySet = wire.NewSet(
 	wire.Bind(new(handleroauth.ChallengeProvider), new(*challenge.Provider)),
 	wire.Bind(new(handleroauth.JSONResponseWriter), new(*httputil.JSONResponseWriter)),
 	wire.Bind(new(handleroauth.AppSessionTokenIssuer), new(*oauthhandler.TokenHandler)),
-	wire.Bind(new(handleroauth.Renderer), new(*handlerwebapp.ResponseRenderer)),
+	wire.Bind(new(handleroauth.Renderer), new(*web.ResponseRenderer)),
 	wire.Bind(new(handleroauth.ProtocolIdentityService), new(*identityservice.Service)),
 	wire.Bind(new(handleroauth.ProtocolProxyRedirectHandler), new(*oauthhandler.ProxyRedirectHandler)),
 	wire.Bind(new(handleroauth.OAuthClientResolver), new(*oauthclient.Resolver)),
