+++ pkg/auth/handler/webapp/deps.go
@@ -6,9 +6,7 @@ var DependencySet = wire.NewSet(
 	wire.Struct(new(AuthEntryPointMiddleware), "*"),
 	wire.Struct(new(ImplementationSwitcherMiddleware), "*"),
 
-	wire.Struct(new(ResponseRenderer), "*"),
 	wire.Struct(new(FormPrefiller), "*"),
-	wire.Bind(new(Renderer), new(*ResponseRenderer)),
 
 	wire.Struct(new(ControllerDeps), "*"),
 	wire.Struct(new(ControllerFactory), "*"),
