+++ pkg/lib/web/deps.go
@@ -6,4 +6,5 @@ import (
 
 var DependencySet = wire.NewSet(
 	wire.Struct(new(StaticAssetResolver), "*"),
+	wire.Struct(new(ResponseRenderer), "*"),
 )
